+++ src/sys/os/linux/time/mod.rs
@@ -9,11 +9,16 @@
 mod clock; // LinuxClock
 mod timespec; // LinuxTimespec
 
+#[cfg(all(feature = "unsafe_syscall", not(miri)))]
+mod instant; // LinuxInstant, LinuxTime
+
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             clock::*,
             timespec::*,
         };
+        #[cfg(all(feature = "unsafe_syscall", not(miri)))]
+        pub use super::instant::*;
     }
 }